assignment_2/requirements.txt
jupyterlab==4.3.5
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
scipy==1.13.1
pyspark==3.5.1
pyarrow==14.0.1

scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.3.0
imbalanced-learn==0.11.0

joblib==1.4.2

apache-airflow==2.9.0
apache-airflow-providers-apache-spark==4.10.0

python-dateutil==2.9.0
pyyaml==6.0.1
tqdm==4.66.4